created messages pages and added styles

+++ app/assets/javascripts/components/babysitter/babysitter-profile-controller.js
@@ -11,38 +11,30 @@ angular.module('app')
       this.userInput = false;
       this.updatedBabysitters = {};
       this.userInput = "";
+      this.changed = true;
       /*************************
       Verifying Certifications
       *************************/
       this.addCerts = function() {
-        $('<div>').attr('class', 'new-cert-container').appendTo('.cert');
-        $('<input>').attr('id', 'cpr').appendTo('.new-cert-container');
-        $('<span>').html('-').appendTo('.new-cert-container');
+
       }
       /*************************
       If user edits input fields, data will be saved in object
       *************************/
       this.updateBabysitter = function(id) {
+        self.changed = true;
         babysitterDirectoryAPI.updateUser(id, self.updatedBabysitters);
       }
       /*************************
       If user deletes babysitter, sitter will be removed from database
       *************************/
       this.deleteBabysitter = function(id) {
-        babysitterDirectoryAPI.deleteUser(id, self.updatedBabysitters);
-        $state.go('babysitters');
+        var result = confirm("Are you sure you want to delete user?");
+        if (result) {
+          babysitterDirectoryAPI.deleteUser(id, self.updatedBabysitters);
+          $state.go('babysitters',{reload: true});
+        }
       }
-      /*************************
-      delete certification
-      *************************/
-      this.deleteCert = function(key, sitter_id, event) {
-        self.updatedBabysitters['id'] = sitter_id;
-        var updatedUser = self.updatedBabysitters[key] = false;
-        babysitterDirectoryAPI.updateUser(sitter_id, self.updatedBabysitters);
-        $(event).remove();
-        console.log(self.updatedBabysitters);
-      }
-
       /*************************
       Convert Strings
       *************************/
